bored lazy bs workaround

+++ src/main/scala/priv/sp/update/PlayerStats.scala
@@ -4,8 +4,8 @@ import priv.sp._
 
 object PlayerStats {
   def basicKillValue(slot : SlotState) = slot.card.cost * (if (slot.card.houseIndex == 4) 1.5f else 1f)
-  def hpManaRatio(slot : SlotState) : Float = hpManaRatio(slot.card)
-  @inline def hpManaRatio(card : Creature) : Float = (card.life / (0.5 + getCostPowMana(card.cost, card.houseIndex))).toFloat
+  def hpManaRatio(slot : SlotState) : Double = hpManaRatio(slot.card)
+  @inline def hpManaRatio(card : Creature): Double = (card.life / (0.5 + getCostPowMana(card.cost, card.houseIndex)))
 
   def getCostMana(m : Int, houseIndex : Int) = {
     if (houseIndex == 4) m * 1.5 else m
@@ -16,8 +16,8 @@ object PlayerStats {
 }
 
 // Some bs for heuristic
-case class PlayerStats(var nbSummon : Int = 0, var nbKill : Int = 0, var nbDead : Int = 0, var killValue : Float = 0){
-  var getKillValue : Function[SlotState, Float] = PlayerStats.hpManaRatio _
+case class PlayerStats(var nbSummon : Int = 0, var nbKill : Int = 0, var nbDead : Int = 0, var killValue : Double = 0){
+  var getKillValue : Function[SlotState, Double] = PlayerStats.hpManaRatio _
 
   def reset(){
     nbSummon = 0
